fixed looping problems and failure of animation
The looping function did not return true so the animation only
ran one time,

The animation stops when the cursor is still for a period of time
and thre result of the Math.sqrt function becomes 0. This generates
a NaN, which in turns causes the animation to halt and not restart
when the cursor moves. The animation will now settle but not break.

NB: Need to introduce small random offsets to reintroduce appearance
that teh squidy is moving all the time.