added params to control tail spacing and postion
This is to tweek where the tails sits when the cursor is stationary. The
default is straight up.

The tail spacing allows you to specify the offset to the tail. This is
handy if the tail image is smaller than than the head. These params
allow the tail to be nudged up closer to the head image for a
more natural effect.